requirements.txt
streamlit
nbformat
openai
python-dotenv